requirements.txt
fastapi==0.111.1
uvicorn==0.30.3
pymongo==4.8.0
pydantic==2.8.2
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
